added support for interactions in shared models
