+++ cs_unittest/cs_unittest.csproj
@@ -62,10 +62,12 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Test3.cs" />
     <Compile Include="Test87.cs" />
+    <Compile Include="TestAntlr.cs" />
     <Compile Include="TestBase.cs" />
     <Compile Include="TestErrorListener.cs" />
     <Compile Include="TestPooling.cs" />
     <Compile Include="TestWrapper.cs" />
+    <Compile Include="TrainSet0002Dat.cs" />
     <Compile Include="VWTestHelper.cs" />
   </ItemGroup>
   <ItemGroup>
